move Debian and Ubuntu to DistributionInstaller
